requirements.txt
absl-py==1.4.0
accelerate==0.16.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.37.0
bleach==6.0.0
blis==0.7.9
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.1
catalogue==2.0.8
cchardet==2.1.7
chardet==5.1.0
click==8.1.3
cmake==3.26.3
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
decord==0.6.0
deepspeed==0.10.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.10
einops==0.6.1
entrypoints==0.4
executing==1.2.0
fastapi==0.95.1
fastjsonschema==2.16.3
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==3.24.1
gradio_client==0.0.8
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
hjson==3.1.0
httpcore==0.17.0
httpx==0.24.0
huggingface-hub==0.13.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.22.0
ipython==8.12.3
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonlines==3.1.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
langcodes==3.3.0
linkify-it-py==2.0.0
lit==16.0.1
llvmlite==0.39.1
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mkl-fft==1.3.1
mkl-service==2.4.0
more-itertools==10.0.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.16.4
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
ninja==1.11.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.2
numba==0.56.4
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.27.0
opencv-python==4.7.0.72
orjson==3.8.10
packaging==23.0
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
peft==0.2.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.5.0
platformdirs==3.2.0
portalocker==2.7.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
pynndescent==0.5.10
pyparsing==3.0.9
pyproject-toml==0.0.10
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2022.10.31
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
-e git+https://github.com/facebookresearch/SEAL.git@fdaf30795982958706882d790f81d2d3545d2406#egg=SEAL
semantic-version==2.10.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.98
sentry-sdk==1.20.0
setproctitle==1.3.2
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.14.0
tensorboard-data-server==0.7.1
terminado==0.17.1
terminaltables==3.1.10
thinc==8.1.9
threadpoolctl==3.1.0
timm==0.6.13
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio
torchvision
tornado==6.3.1
tqdm==4.64.1
traitlets==5.9.0
transformers==4.28.0
triton==2.0.0
typer==0.7.0
typing_extensions
tzdata==2023.3
uc-micro-py==1.0.1
umap-learn==0.5.3
uri-template==1.2.0
urllib3
uvicorn==0.21.1
wandb==0.15.0
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.3.7
yarg==0.1.9
yarl==1.8.2
zipp==3.14.0
